create reverse proxy routing
big changes here.  now not only does it auto route to gameservers, but also when gameservers go down they are taken off the list.  when the upgrade header is set, it doesn't do anything yet.  I'm a bit worried about those cookies since I'm messing with routing, but if they are a problem I can probably just clear that