set up HUD
going to make everything windowed.  need to think of a good way to organize localization...  everything is going to have a logout function, which is guaranteed to reset everything to its original state without screwing up the DOM.  I'd hate to have to clear the body and then reinitialize game just to have someone log out.