increment version to 0.1.1
here it is, folks!  by that I mean my one-person audience.  anyway 0.1.1 has a ton of changes, espectially, especially server side. don't ask me what the official release of 0.1.0 was because it doesn't exist.  I've just been using 0.1.0 until now.  technically, the first commit where the version was 0.1.0 was that initial commit, so these changes are a pretty big deal.  anyway, it incorporates everything from custom element localization with jquery ui to abstracted websocket communication behind a reverse proxy to a fancy load balancing system on the backend.  but where is the game, you may ask!  why is the renderer still the same spinning crate?  well, that's classified information :DDDDD