fix ents reassign bug, add broadcast system
I'm becoming worried that at this stage my code has turned to/is going to turn to spaghetti, so I'm trying to decouple things as much as possible.  also, I really want to have users be like 'controllers' for characters, and not be strictly associated with them.  it makes AI easier, and makes code easier in the long run.  I just need to do it right...